+++ CHANGES.md
@@ -3,3 +3,4 @@
 ## 0.1.1
 
 * support `null` value in hocon
+* added feature `serde-support` to allow deserializing HOCON documents to a struct, enabled by default